feat: introduce `ExtensionNavigation` component and streamline navigation logic
- Added `ExtensionNavigation` and replaced `DashboardNavigation` in `ExtensionWindowHeader`.
- Extracted navigation logic into a reusable `useNavigation` composable for improved modularity.
- Updated imports and consolidated navigation-related types for consistency.
